pyproject.toml
[project]
name = "mca-routing-bscap"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asteval>=1.0.8",
    "cython>=3.2.4",
    "jsonschema>=4.26.0",
    "line-profiler>=5.0.0",
    "matplotlib>=3.10.8",
    "mypy>=1.19.1",
    "networkx>=3.6.1",
    "nuitka>=2.8.9",
    "numba>=0.63.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "prompt-toolkit>=3.0.52",
    "psutil>=7.2.1",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "simpy>=4.1.1",
    "ttkbootstrap>=1.20.1",
    "types-networkx>=3.6.1.20251220",
]

[dependency-groups]
dev = []
